# chunkIterator: optimize caching and AtT() calculation